UserInfo -> Schema 수정 , 배열이 아닌데 배열로 선언해놓은것, int64(integer) -> Java에선 long 형으로 수정

+++ src/main/java/com/example/vraapi/identity/Schemas/Role.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class Role {
     private String id;
-    private int createdMillis;
-    private int updatedMillis;
+    private long createdMillis;
+    private long updatedMillis;
     private String name;
     private String displayName;
     private String orgId;
